@@ -23,14 +23,13 @@ class CandleStick extends Bar {
     this.series = series
     this.yRatio = coreUtils.getLogYRatios(this.yRatio)
 
-    this.initVariables(series)
+    this.barHelpers.initVariables(series)
 
     let ret = graphics.group({
       class: 'apexcharts-candlestick-series apexcharts-plot-series'
     })
 
-    for (let i = 0, bc = 0; i < series.length; i++, bc++) {
-      let pathTo, pathFrom
+    for (let i = 0; i < series.length; i++) {
       let x,
         y,
         xDivision, // xDivision is the GRIDWIDTH divided by number of datapoints (columns)
@@ -53,7 +52,6 @@ class CandleStick extends Bar {
         this.visibleI = this.visibleI + 1
       }
 
-      let strokeWidth = 0
       let barHeight = 0
       let barWidth = 0
 
@@ -61,7 +59,7 @@ class CandleStick extends Bar {
         this.yaxisIndex = realIndex
       }
 
-      let initPositions = this.initialPositions()
+      let initPositions = this.barHelpers.initialPositions()
 
       y = initPositions.y
       barHeight = initPositions.barHeight
@@ -78,25 +76,8 @@ class CandleStick extends Bar {
         class: 'apexcharts-datalabels'
       })
 
-      for (
-        let j = 0, tj = w.globals.dataPoints;
-        j < w.globals.dataPoints;
-        j++, tj--
-      ) {
-        if (typeof this.series[i][j] === 'undefined' || series[i][j] === null) {
-          this.isNullValue = true
-        } else {
-          this.isNullValue = false
-        }
-        if (w.config.stroke.show) {
-          if (this.isNullValue) {
-            strokeWidth = 0
-          } else {
-            strokeWidth = Array.isArray(this.strokeWidth)
-              ? this.strokeWidth[realIndex]
-              : this.strokeWidth
-          }
-        }
+      for (let j = 0; j < w.globals.dataPoints; j++) {
+        const strokeWidth = this.barHelpers.getStrokeWidth(i, j, realIndex)
 
         let color
 
@@ -104,22 +85,17 @@ class CandleStick extends Bar {
           indexes: {
             i,
             j,
-            realIndex,
-            bc
+            realIndex
           },
           x,
           y,
           xDivision,
-          pathTo,
-          pathFrom,
           barWidth,
           zeroH,
           strokeWidth,
           elSeries
         })
 
-        pathTo = paths.pathTo
-        pathFrom = paths.pathFrom
         y = paths.y
         x = paths.x
         color = paths.color
@@ -133,6 +109,7 @@ class CandleStick extends Bar {
 
         let pathFill = fill.fillPath({
           seriesNumber: realIndex,
+          dataPointIndex: j,
           color,
           value: series[i][j]
         })
@@ -141,14 +118,14 @@ class CandleStick extends Bar {
           ? color
           : undefined
 
-        elSeries = this.renderSeries({
+        this.renderSeries({
           realIndex,
           pathFill,
           lineFill,
           j,
           i,
-          pathFrom,
-          pathTo,
+          pathFrom: paths.pathFrom,
+          pathTo: paths.pathTo,
           strokeWidth,
           elSeries,
           x,
@@ -177,8 +154,6 @@ class CandleStick extends Bar {
     x,
     y,
     xDivision,
-    pathTo,
-    pathFrom,
     barWidth,
     zeroH,
     strokeWidth
@@ -209,7 +184,8 @@ class CandleStick extends Bar {
 
     if (w.globals.isXNumeric) {
       x =
-        (w.globals.seriesX[i][j] - w.globals.minX) / this.xRatio - barWidth / 2
+        (w.globals.seriesX[realIndex][j] - w.globals.minX) / this.xRatio -
+        barWidth / 2
     }
 
     let barXPosition = x + barWidth * this.visibleI
@@ -226,10 +202,10 @@ class CandleStick extends Bar {
       l2 = zeroH - ohlc.l / yRatio
     }
 
-    pathTo = graphics.move(barXPosition, zeroH)
-    pathFrom = graphics.move(barXPosition, y1)
+    let pathTo = graphics.move(barXPosition, zeroH)
+    let pathFrom = graphics.move(barXPosition, y1)
     if (w.globals.previousPaths.length > 0) {
-      pathFrom = this.getPathFrom(realIndex, j, true)
+      pathFrom = this.getPreviousPath(realIndex, j, true)
     }
 
     pathTo =
