@@ -20,8 +20,8 @@ class Fill {
     let w = this.w
     let cnf = w.config
 
-    let svgW = parseInt(w.globals.gridWidth)
-    let svgH = parseInt(w.globals.gridHeight)
+    let svgW = parseInt(w.globals.gridWidth, 10)
+    let svgH = parseInt(w.globals.gridHeight, 10)
 
     let size = svgW > svgH ? svgW : svgH
 
@@ -81,7 +81,8 @@ class Fill {
     const w = this.w
 
     if (
-      (w.config.chart.type === 'bar' && w.config.plotOptions.bar.distributed) ||
+      ((w.config.chart.type === 'bar' || w.config.chart.type === 'rangeBar') &&
+        w.config.plotOptions.bar.distributed) ||
       w.config.chart.type === 'heatmap'
     ) {
       this.seriesIndex = opts.seriesNumber
@@ -109,6 +110,7 @@ class Fill {
     if (typeof fillColor === 'function') {
       fillColor = fillColor({
         seriesIndex: this.seriesIndex,
+        dataPointIndex: opts.dataPointIndex,
         value: opts.value,
         w
       })
@@ -131,6 +133,7 @@ class Fill {
         fillOpacity = 0 + '.' + Utils.getOpacityFromRGBA(fillColor)
       }
     }
+    if (opts.opacity) fillOpacity = opts.opacity
 
     if (fillType === 'pattern') {
       patternFill = this.handlePatternFill(
@@ -150,18 +153,25 @@ class Fill {
       )
     }
 
-    if (cnf.fill.image.src.length > 0 && fillType === 'image') {
-      if (opts.seriesNumber < cnf.fill.image.src.length) {
-        this.clippedImgArea({
-          opacity: fillOpacity,
-          image: cnf.fill.image.src[opts.seriesNumber],
-          patternUnits: opts.patternUnits,
-          patternID: `pattern${w.globals.cuid}${opts.seriesNumber + 1}`
-        })
-        pathFill = `url(#pattern${w.globals.cuid}${opts.seriesNumber + 1})`
-      } else {
-        pathFill = defaultColor
-      }
+    if (fillType === 'image') {
+      let imgSrc = cnf.fill.image.src
+
+      let patternID = opts.patternID ? opts.patternID : ''
+      this.clippedImgArea({
+        opacity: fillOpacity,
+        image: Array.isArray(imgSrc)
+          ? opts.seriesNumber < imgSrc.length
+            ? imgSrc[opts.seriesNumber]
+            : imgSrc[0]
+          : imgSrc,
+        width: opts.width ? opts.width : undefined,
+        height: opts.height ? opts.height : undefined,
+        patternUnits: opts.patternUnits,
+        patternID: `pattern${w.globals.cuid}${opts.seriesNumber +
+          1}${patternID}`
+      })
+      pathFill = `url(#pattern${w.globals.cuid}${opts.seriesNumber +
+        1}${patternID})`
     } else if (fillType === 'gradient') {
       pathFill = gradientFill
     } else if (fillType === 'pattern') {
