@@ -8,12 +8,13 @@ export default class Range {
 
   // http://stackoverflow.com/questions/326679/choosing-an-attractive-linear-scale-for-a-graphs-y-axiss
   // This routine creates the Y axis values for a graph.
-  niceScale(yMin, yMax, diff, index = 0, ticks = 10) {
+  niceScale(yMin, yMax, diff, index = 0, ticks = 10, NO_MIN_MAX_PROVIDED) {
     const w = this.w
-    const NO_MIN_MAX_PROVIDED =
-      (this.w.config.yaxis[index].max === undefined &&
-        this.w.config.yaxis[index].min === undefined) ||
-      this.w.config.yaxis[index].forceNiceScale
+
+    if (ticks === 'dataPoints') {
+      ticks = w.globals.dataPoints - 1
+    }
+
     if (
       (yMin === Number.MIN_VALUE && yMax === 0) ||
       (!Utils.isNumber(yMin) && !Utils.isNumber(yMax)) ||
@@ -29,7 +30,7 @@ export default class Range {
     if (yMin > yMax) {
       // if somehow due to some wrong config, user sent max less than min,
       // adjust the min/max again
-      console.warn('yaxis.min cannot be greater than yaxis.max')
+      console.warn('axis.min cannot be greater than axis.max')
       yMax = yMin + 0.1
     } else if (yMin === yMax) {
       // If yMin and yMax are identical, then
@@ -114,7 +115,7 @@ export default class Range {
         result.push(v)
       }
 
-      if (result[result.length - 2] >= yMax) {
+      if (result[result.length - 1] >= yMax) {
         result.pop()
       }
 
@@ -178,7 +179,7 @@ export default class Range {
       }
 
       // calculate adjustment factor
-      var scale = (max - min) / (yMax - yMin)
+      let scale = (max - min) / (yMax - yMin)
 
       const logVal = Math.pow(base, min + scale * (niceNumber - min))
       return (
@@ -207,7 +208,13 @@ export default class Range {
       gl.yAxisScale[index] = []
     }
 
-    if (y.logarithmic) {
+    let diff = Math.abs(maxY - minY)
+
+    if (y.logarithmic && diff <= 5) {
+      gl.invalidLogScale = true
+    }
+
+    if (y.logarithmic && diff > 5) {
       gl.allSeriesCollapsed = false
       gl.yAxisScale[index] = this.logarithmicScale(
         index,
@@ -227,19 +234,42 @@ export default class Range {
           // fix https://github.com/apexcharts/apexcharts.js/issues/492
           gl.yAxisScale[index] = this.linearScale(minY, maxY, y.tickAmount)
         } else {
-          let diff = Math.abs(maxY - minY)
-
+          const noMinMaxProvided =
+            (cnf.yaxis[index].max === undefined &&
+              cnf.yaxis[index].min === undefined) ||
+            cnf.yaxis[index].forceNiceScale
           gl.yAxisScale[index] = this.niceScale(
             minY,
             maxY,
             diff,
             index,
             // fix https://github.com/apexcharts/apexcharts.js/issues/397
-            y.tickAmount ? y.tickAmount : diff < 5 && diff > 1 ? diff + 1 : 5
+            y.tickAmount ? y.tickAmount : diff < 5 && diff > 1 ? diff + 1 : 5,
+            noMinMaxProvided
           )
         }
       }
     }
+  }
+
+  setXScale(minX, maxX) {
+    const w = this.w
+    const gl = w.globals
+    const x = w.config.xaxis
+    let diff = Math.abs(maxX - minX)
+    if (maxX === -Number.MAX_VALUE || !Utils.isNumber(maxX)) {
+      // no data in the chart. Either all series collapsed or user passed a blank array
+      gl.xAxisScale = this.linearScale(0, 5, 5)
+    } else {
+      gl.xAxisScale = this.niceScale(
+        minX,
+        maxX,
+        diff,
+        0,
+        x.tickAmount ? x.tickAmount : diff < 5 && diff > 1 ? diff + 1 : 5
+      )
+    }
+    return gl.xAxisScale
   }
 
   setMultipleYScales() {
@@ -317,29 +347,21 @@ export default class Range {
     })
 
     // then, we remove duplicates from the similarScale array
-    let uniqueSimilarIndices = similarIndices.map(function(item) {
-      return item.filter((i, pos) => {
-        return item.indexOf(i) === pos
-      })
+    let uniqueSimilarIndices = similarIndices.map((item) => {
+      return item.filter((i, pos) => item.indexOf(i) === pos)
     })
 
     // sort further to remove whole duplicate arrays later
-    let sortedIndices = uniqueSimilarIndices.map((s) => {
-      return s.sort()
-    })
+    let sortedIndices = uniqueSimilarIndices.map((s) => s.sort())
 
     // remove undefined items
-    similarIndices = similarIndices.filter((s) => {
-      return !!s
-    })
+    similarIndices = similarIndices.filter((s) => !!s)
 
     let indices = sortedIndices.slice()
-    let stringIndices = indices.map((ind) => {
-      return JSON.stringify(ind)
-    })
-    indices = indices.filter((ind, p) => {
-      return stringIndices.indexOf(JSON.stringify(ind)) === p
-    })
+    let stringIndices = indices.map((ind) => JSON.stringify(ind))
+    indices = indices.filter(
+      (ind, p) => stringIndices.indexOf(JSON.stringify(ind)) === p
+    )
 
     let sameScaleMinYArr = []
     let sameScaleMaxYArr = []
@@ -435,6 +457,7 @@ export default class Range {
     if (w.globals.isMultipleYAxis || w.globals.collapsedSeries.length) {
       // The autoScale option for multiple y-axis is turned off as it leads to buggy behavior.
       // Also, when a series is collapsed, it results in incorrect behavior. Hence turned it off for that too - fixes apexcharts.js#795
+      console.warn('autoScaleYaxis is not supported in a multi-yaxis chart.')
       return yaxis
     }
 
@@ -442,7 +465,7 @@ export default class Range {
 
     let isStacked = w.config.chart.stacked
 
-    yaxis.forEach((yaxe, yI) => {
+    yaxis.forEach((yaxe, yi) => {
       let firstXIndex = 0
 
       for (let xi = 0; xi < seriesX.length; xi++) {
@@ -452,8 +475,8 @@ export default class Range {
         }
       }
 
-      let initialMin = w.globals.minYArr[yI]
-      let initialMax = w.globals.maxYArr[yI]
+      let initialMin = w.globals.minYArr[yi]
+      let initialMax = w.globals.maxYArr[yi]
       let min, max
 
       let stackedSer = w.globals.stackedSeriesTotals
