@@ -9,8 +9,14 @@ import Utils from '../utils/Utils'
 export default class Theme {
   constructor(ctx) {
     this.ctx = ctx
-    this.w = ctx.w
     this.colors = []
+    this.w = ctx.w
+    const w = this.w
+
+    this.isColorFn = false
+    this.isBarDistributed =
+      w.config.plotOptions.bar.distributed &&
+      (w.config.chart.type === 'bar' || w.config.chart.type === 'rangeBar')
   }
 
   init() {
@@ -21,21 +27,47 @@ export default class Theme {
     let w = this.w
     let utils = new Utils()
 
-    w.globals.dom.elWrap.classList.add(w.config.theme.mode)
+    w.globals.dom.elWrap.classList.add(
+      `apexcharts-theme-${w.config.theme.mode}`
+    )
 
     if (w.config.colors === undefined) {
       w.globals.colors = this.predefined()
     } else {
       w.globals.colors = w.config.colors
+
+      // if user provided a function in colors, we need to eval here
+      if (
+        w.globals.axisCharts &&
+        w.config.chart.type !== 'bar' &&
+        Array.isArray(w.config.colors) &&
+        w.config.colors.length > 0 &&
+        w.config.colors.length === w.config.series.length
+        // colors & series length needs same
+      ) {
+        w.globals.colors = w.config.colors.map((c, i) => {
+          if (typeof c === 'function') {
+            this.isColorFn = true
+            return c({
+              value: w.globals.axisCharts
+                ? w.globals.series[i][0]
+                  ? w.globals.series[i][0]
+                  : 0
+                : w.globals.series[i],
+              seriesIndex: i,
+              dataPointIndex: i,
+              w
+            })
+          }
+          return c
+        })
+      }
     }
 
     if (w.config.theme.monochrome.enabled) {
       let monoArr = []
       let glsCnt = w.globals.series.length
-      if (
-        w.config.plotOptions.bar.distributed &&
-        w.config.chart.type === 'bar'
-      ) {
+      if (this.isBarDistributed) {
         glsCnt = w.globals.series[0].length * w.globals.series.length
       }
 
@@ -66,7 +98,7 @@ export default class Theme {
 
     // The Border colors
     if (w.config.stroke.colors === undefined) {
-      w.globals.stroke.colors = defaultColors
+      w.globals.stroke.colors = this.isColorFn ? w.config.colors : defaultColors
     } else {
       w.globals.stroke.colors = w.config.stroke.colors
     }
@@ -74,7 +106,7 @@ export default class Theme {
 
     // The FILL colors
     if (w.config.fill.colors === undefined) {
-      w.globals.fill.colors = defaultColors
+      w.globals.fill.colors = this.isColorFn ? w.config.colors : defaultColors
     } else {
       w.globals.fill.colors = w.config.fill.colors
     }
@@ -117,8 +149,7 @@ export default class Theme {
 
     if (distributed === null) {
       distributed =
-        (w.config.chart.type === 'bar' &&
-          w.config.plotOptions.bar.distributed) ||
+        this.isBarDistributed ||
         (w.config.chart.type === 'heatmap' &&
           w.config.plotOptions.heatmap.colorScale.inverse)
     }
