@@ -22,92 +22,91 @@ export default class Toolbar {
     this.w = ctx.w
 
     this.ev = this.w.config.chart.events
+    this.selectedClass = 'apexcharts-selected'
 
     this.localeValues = this.w.globals.locale.toolbar
   }
 
   createToolbar() {
     let w = this.w
-    const elToolbarWrap = document.createElement('div')
+
+    const createDiv = () => {
+      return document.createElement('div')
+    }
+    const elToolbarWrap = createDiv()
     elToolbarWrap.setAttribute('class', 'apexcharts-toolbar')
     w.globals.dom.elWrap.appendChild(elToolbarWrap)
 
-    this.elZoom = document.createElement('div')
-    this.elZoomIn = document.createElement('div')
-    this.elZoomOut = document.createElement('div')
-    this.elPan = document.createElement('div')
-    this.elSelection = document.createElement('div')
-    this.elZoomReset = document.createElement('div')
-    this.elMenuIcon = document.createElement('div')
-    this.elMenu = document.createElement('div')
+    this.elZoom = createDiv()
+    this.elZoomIn = createDiv()
+    this.elZoomOut = createDiv()
+    this.elPan = createDiv()
+    this.elSelection = createDiv()
+    this.elZoomReset = createDiv()
+    this.elMenuIcon = createDiv()
+    this.elMenu = createDiv()
     this.elCustomIcons = []
 
     this.t = w.config.chart.toolbar.tools
 
     if (Array.isArray(this.t.customIcons)) {
       for (let i = 0; i < this.t.customIcons.length; i++) {
-        this.elCustomIcons.push(document.createElement('div'))
+        this.elCustomIcons.push(createDiv())
       }
     }
 
-    this.elMenuItems = []
-
     let toolbarControls = []
 
-    if (this.t.zoomin && w.config.chart.zoom.enabled) {
-      toolbarControls.push({
-        el: this.elZoomIn,
-        icon: typeof this.t.zoomin === 'string' ? this.t.zoomin : icoZoomIn,
-        title: this.localeValues.zoomIn,
-        class: 'apexcharts-zoom-in-icon'
-      })
+    const appendZoomControl = (type, el, ico) => {
+      const tool = type.toLowerCase()
+      if (this.t[tool] && w.config.chart.zoom.enabled) {
+        toolbarControls.push({
+          el,
+          icon: typeof this.t[tool] === 'string' ? this.t[tool] : ico,
+          title: this.localeValues[type],
+          class: `apexcharts-${tool}-icon`
+        })
+      }
     }
 
-    if (this.t.zoomout && w.config.chart.zoom.enabled) {
-      toolbarControls.push({
-        el: this.elZoomOut,
-        icon: typeof this.t.zoomout === 'string' ? this.t.zoomout : icoZoomOut,
-        title: this.localeValues.zoomOut,
-        class: 'apexcharts-zoom-out-icon'
-      })
-    }
+    appendZoomControl('zoomIn', this.elZoomIn, icoZoomIn)
+    appendZoomControl('zoomOut', this.elZoomOut, icoZoomOut)
 
-    if (this.t.zoom && w.config.chart.zoom.enabled) {
-      toolbarControls.push({
-        el: this.elZoom,
-        icon: typeof this.t.zoom === 'string' ? this.t.zoom : icoZoom,
-        title: this.localeValues.selectionZoom,
-        class: w.globals.isTouchDevice ? 'hidden' : 'apexcharts-zoom-icon'
-      })
+    const zoomSelectionCtrls = (z) => {
+      if (this.t[z] && w.config.chart[z].enabled) {
+        toolbarControls.push({
+          el: z === 'zoom' ? this.elZoom : this.elSelection,
+          icon:
+            typeof this.t[z] === 'string'
+              ? this.t[z]
+              : z === 'zoom'
+              ? icoZoom
+              : icoSelect,
+          title: this.localeValues[
+            z === 'zoom' ? 'selectionZoom' : 'selection'
+          ],
+          class: w.globals.isTouchDevice
+            ? 'apexcharts-element-hidden'
+            : `apexcharts-${z}-icon`
+        })
+      }
     }
-
-    if (this.t.selection && w.config.chart.selection.enabled) {
-      toolbarControls.push({
-        el: this.elSelection,
-        icon:
-          typeof this.t.selection === 'string' ? this.t.selection : icoSelect,
-        title: this.localeValues.selection,
-        class: w.globals.isTouchDevice ? 'hidden' : 'apexcharts-selection-icon'
-      })
-    }
+    zoomSelectionCtrls('zoom')
+    zoomSelectionCtrls('selection')
 
     if (this.t.pan && w.config.chart.zoom.enabled) {
       toolbarControls.push({
         el: this.elPan,
         icon: typeof this.t.pan === 'string' ? this.t.pan : icoPan,
         title: this.localeValues.pan,
-        class: w.globals.isTouchDevice ? 'hidden' : 'apexcharts-pan-icon'
+        class: w.globals.isTouchDevice
+          ? 'apexcharts-element-hidden'
+          : 'apexcharts-pan-icon'
       })
     }
 
-    if (this.t.reset && w.config.chart.zoom.enabled) {
-      toolbarControls.push({
-        el: this.elZoomReset,
-        icon: typeof this.t.reset === 'string' ? this.t.reset : icoReset,
-        title: this.localeValues.reset,
-        class: 'apexcharts-reset-zoom-icon'
-      })
-    }
+    appendZoomControl('reset', this.elZoomReset, icoReset)
+
     if (this.t.download) {
       toolbarControls.push({
         el: this.elMenuIcon,
@@ -143,7 +142,22 @@ export default class Toolbar {
       elToolbarWrap.appendChild(toolbarControls[i].el)
     }
 
-    elToolbarWrap.appendChild(this.elMenu)
+    this._createHamburgerMenu(elToolbarWrap)
+
+    if (w.globals.zoomEnabled) {
+      this.elZoom.classList.add(this.selectedClass)
+    } else if (w.globals.panEnabled) {
+      this.elPan.classList.add(this.selectedClass)
+    } else if (w.globals.selectionEnabled) {
+      this.elSelection.classList.add(this.selectedClass)
+    }
+
+    this.addToolbarEventListeners()
+  }
+
+  _createHamburgerMenu(parent) {
+    this.elMenuItems = []
+    parent.appendChild(this.elMenu)
 
     Graphics.setAttrs(this.elMenu, {
       class: 'apexcharts-menu'
@@ -157,6 +171,10 @@ export default class Toolbar {
       {
         name: 'exportPNG',
         title: this.localeValues.exportToPNG
+      },
+      {
+        name: 'exportCSV',
+        title: this.localeValues.exportToCSV
       }
     ]
     for (let i = 0; i < menuItems.length; i++) {
@@ -168,31 +186,29 @@ export default class Toolbar {
       })
       this.elMenu.appendChild(this.elMenuItems[i])
     }
-
-    if (w.globals.zoomEnabled) {
-      this.elZoom.classList.add('selected')
-    } else if (w.globals.panEnabled) {
-      this.elPan.classList.add('selected')
-    } else if (w.globals.selectionEnabled) {
-      this.elSelection.classList.add('selected')
-    }
-
-    this.addToolbarEventListeners()
   }
 
   addToolbarEventListeners() {
     this.elZoomReset.addEventListener('click', this.handleZoomReset.bind(this))
-    this.elSelection.addEventListener('click', this.toggleSelection.bind(this))
-    this.elZoom.addEventListener('click', this.toggleZooming.bind(this))
+    this.elSelection.addEventListener(
+      'click',
+      this.toggleZoomSelection.bind(this, 'selection')
+    )
+    this.elZoom.addEventListener(
+      'click',
+      this.toggleZoomSelection.bind(this, 'zoom')
+    )
     this.elZoomIn.addEventListener('click', this.handleZoomIn.bind(this))
     this.elZoomOut.addEventListener('click', this.handleZoomOut.bind(this))
     this.elPan.addEventListener('click', this.togglePanning.bind(this))
     this.elMenuIcon.addEventListener('click', this.toggleMenu.bind(this))
     this.elMenuItems.forEach((m) => {
       if (m.classList.contains('exportSVG')) {
-        m.addEventListener('click', this.downloadSVG.bind(this))
+        m.addEventListener('click', this.handleDownload.bind(this, 'svg'))
       } else if (m.classList.contains('exportPNG')) {
-        m.addEventListener('click', this.downloadPNG.bind(this))
+        m.addEventListener('click', this.handleDownload.bind(this, 'png'))
+      } else if (m.classList.contains('exportCSV')) {
+        m.addEventListener('click', this.handleDownload.bind(this, 'csv'))
       }
     })
     for (let i = 0; i < this.t.customIcons.length; i++) {
@@ -203,25 +219,18 @@ export default class Toolbar {
     }
   }
 
-  toggleSelection() {
+  toggleZoomSelection(type) {
     this.toggleOtherControls()
-    this.w.globals.selectionEnabled = !this.w.globals.selectionEnabled
 
-    if (!this.elSelection.classList.contains('selected')) {
-      this.elSelection.classList.add('selected')
+    let el = type === 'selection' ? this.elSelection : this.elZoom
+    let enabledType = type === 'selection' ? 'selectionEnabled' : 'zoomEnabled'
+
+    this.w.globals[enabledType] = !this.w.globals[enabledType]
+
+    if (!el.classList.contains(this.selectedClass)) {
+      el.classList.add(this.selectedClass)
     } else {
-      this.elSelection.classList.remove('selected')
-    }
-  }
-
-  toggleZooming() {
-    this.toggleOtherControls()
-    this.w.globals.zoomEnabled = !this.w.globals.zoomEnabled
-
-    if (!this.elZoom.classList.contains('selected')) {
-      this.elZoom.classList.add('selected')
-    } else {
-      this.elZoom.classList.remove('selected')
+      el.classList.remove(this.selectedClass)
     }
   }
 
@@ -240,26 +249,20 @@ export default class Toolbar {
     }
   }
 
-  enableZooming() {
+  enableZoomPanFromToolbar(type) {
     this.toggleOtherControls()
-    this.w.globals.zoomEnabled = true
-    if (this.elZoom) {
-      this.elZoom.classList.add('selected')
-    }
-    if (this.elPan) {
-      this.elPan.classList.remove('selected')
-    }
-  }
 
-  enablePanning() {
-    this.toggleOtherControls()
-    this.w.globals.panEnabled = true
+    type === 'pan'
+      ? (this.w.globals.panEnabled = true)
+      : (this.w.globals.zoomEnabled = true)
 
-    if (this.elPan) {
-      this.elPan.classList.add('selected')
+    const el = type === 'pan' ? this.elPan : this.elZoom
+    const el2 = type === 'pan' ? this.elZoom : this.elPan
+    if (el) {
+      el.classList.add(this.selectedClass)
     }
-    if (this.elZoom) {
-      this.elZoom.classList.remove('selected')
+    if (el2) {
+      el2.classList.remove(this.selectedClass)
     }
   }
 
@@ -267,10 +270,10 @@ export default class Toolbar {
     this.toggleOtherControls()
     this.w.globals.panEnabled = !this.w.globals.panEnabled
 
-    if (!this.elPan.classList.contains('selected')) {
-      this.elPan.classList.add('selected')
+    if (!this.elPan.classList.contains(this.selectedClass)) {
+      this.elPan.classList.add(this.selectedClass)
     } else {
-      this.elPan.classList.remove('selected')
+      this.elPan.classList.remove(this.selectedClass)
     }
   }
 
@@ -282,26 +285,25 @@ export default class Toolbar {
 
     this.getToolbarIconsReference()
 
-    if (this.elPan) {
-      this.elPan.classList.remove('selected')
-    }
-    if (this.elSelection) {
-      this.elSelection.classList.remove('selected')
-    }
-    if (this.elZoom) {
-      this.elZoom.classList.remove('selected')
-    }
+    const toggleEls = [this.elPan, this.elSelection, this.elZoom]
+    toggleEls.forEach((el) => {
+      if (el) {
+        el.classList.remove(this.selectedClass)
+      }
+    })
   }
 
   handleZoomIn() {
     const w = this.w
 
     const centerX = (w.globals.minX + w.globals.maxX) / 2
-    const newMinX = (w.globals.minX + centerX) / 2
-    const newMaxX = (w.globals.maxX + centerX) / 2
+    let newMinX = (w.globals.minX + centerX) / 2
+    let newMaxX = (w.globals.maxX + centerX) / 2
+
+    const newMinXMaxX = this._getNewMinXMaxX(newMinX, newMaxX)
 
     if (!w.globals.disableZoomIn) {
-      this.zoomUpdateOptions(newMinX, newMaxX)
+      this.zoomUpdateOptions(newMinXMaxX.minX, newMinXMaxX.maxX)
     }
   }
 
@@ -317,16 +319,38 @@ export default class Toolbar {
     }
 
     const centerX = (w.globals.minX + w.globals.maxX) / 2
-    const newMinX = w.globals.minX - (centerX - w.globals.minX)
-    const newMaxX = w.globals.maxX - (centerX - w.globals.maxX)
+    let newMinX = w.globals.minX - (centerX - w.globals.minX)
+    let newMaxX = w.globals.maxX - (centerX - w.globals.maxX)
+
+    const newMinXMaxX = this._getNewMinXMaxX(newMinX, newMaxX)
 
     if (!w.globals.disableZoomOut) {
-      this.zoomUpdateOptions(newMinX, newMaxX)
+      this.zoomUpdateOptions(newMinXMaxX.minX, newMinXMaxX.maxX)
+    }
+  }
+
+  _getNewMinXMaxX(newMinX, newMaxX) {
+    const shouldFloor = this.w.config.xaxis.convertedCatToNumeric
+    return {
+      minX: shouldFloor ? Math.floor(newMinX) : newMinX,
+      maxX: shouldFloor ? Math.floor(newMaxX) : newMaxX
     }
   }
 
   zoomUpdateOptions(newMinX, newMaxX) {
     const w = this.w
+
+    if (w.config.xaxis.convertedCatToNumeric) {
+      // in category charts, avoid zooming out beyond min and max
+      if (newMinX < 1) {
+        newMinX = 1
+        newMaxX = w.globals.dataPoints
+      }
+
+      if (newMaxX - newMinX < 2) {
+        return
+      }
+    }
 
     let xaxis = {
       min: newMinX,
@@ -358,7 +382,7 @@ export default class Toolbar {
 
     this.w.globals.zoomed = true
 
-    this.ctx._updateOptions(
+    this.ctx.updateHelpers._updateOptions(
       options,
       false,
       this.w.config.chart.animations.dynamicAnimation.enabled
@@ -383,23 +407,29 @@ export default class Toolbar {
   }
 
   toggleMenu() {
-    if (this.elMenu.classList.contains('open')) {
-      this.elMenu.classList.remove('open')
-    } else {
-      this.elMenu.classList.add('open')
+    window.setTimeout(() => {
+      if (this.elMenu.classList.contains('apexcharts-menu-open')) {
+        this.elMenu.classList.remove('apexcharts-menu-open')
+      } else {
+        this.elMenu.classList.add('apexcharts-menu-open')
+      }
+    }, 0)
+  }
+
+  handleDownload(type) {
+    const w = this.w
+    const exprt = new Exports(this.ctx)
+    switch (type) {
+      case 'svg':
+        exprt.exportToSVG(this.ctx)
+        break
+      case 'png':
+        exprt.exportToPng(this.ctx)
+        break
+      case 'csv':
+        exprt.exportToCSV({ series: w.config.series })
+        break
     }
-  }
-
-  downloadPNG() {
-    const downloadPNG = new Exports(this.ctx)
-    downloadPNG.exportToPng(this.ctx)
-    this.toggleMenu()
-  }
-
-  downloadSVG() {
-    const downloadSVG = new Exports(this.ctx)
-    downloadSVG.exportToSVG()
-    this.toggleMenu()
   }
 
   handleZoomReset(e) {
@@ -409,10 +439,10 @@ export default class Toolbar {
       let w = ch.w
 
       if (
-        w.globals.minX !== w.globals.initialminX &&
-        w.globals.maxX !== w.globals.initialmaxX
+        w.globals.minX !== w.globals.initialMinX ||
+        w.globals.maxX !== w.globals.initialMaxX
       ) {
-        ch.revertDefaultAxisMinMax()
+        ch.updateHelpers.revertDefaultAxisMinMax()
 
         if (typeof w.config.chart.events.zoomed === 'function') {
           this.zoomCallback({
@@ -423,7 +453,7 @@ export default class Toolbar {
 
         w.globals.zoomed = false
 
-        ch._updateSeries(
+        ch.updateHelpers._updateSeries(
           w.globals.initialSeries,
           w.config.chart.animations.dynamicAnimation.enabled
         )
