@@ -37,11 +37,12 @@ export default class ZoomPanSelection extends Toolbar {
     this.startY = 0
     this.endY = 0
     this.dragY = 0
+    this.moveDirection = 'none'
   }
 
   init({ xyRatios }) {
     let w = this.w
-    var me = this
+    let me = this
 
     this.xyRatios = xyRatios
 
@@ -78,7 +79,7 @@ export default class ZoomPanSelection extends Toolbar {
     this.preselectedSelection()
 
     this.hoverArea = w.globals.dom.baseEl.querySelector(w.globals.chartClass)
-    this.hoverArea.classList.add('zoomable')
+    this.hoverArea.classList.add('apexcharts-zoomable')
 
     this.eventList.forEach((event) => {
       this.hoverArea.addEventListener(
@@ -116,10 +117,10 @@ export default class ZoomPanSelection extends Toolbar {
 
     if (e.shiftKey) {
       this.shiftWasPressed = true
-      toolbar.enablePanning()
+      toolbar.enableZoomPanFromToolbar('pan')
     } else {
       if (this.shiftWasPressed) {
-        toolbar.enableZooming()
+        toolbar.enableZoomPanFromToolbar('zoom')
         this.shiftWasPressed = false
       }
     }
@@ -201,6 +202,9 @@ export default class ZoomPanSelection extends Toolbar {
         }
       }
 
+      if (w.globals.panEnabled) {
+        me.delayedPanScrolled()
+      }
       if (w.globals.zoomEnabled) {
         me.hideSelectionRect(this.selectionRect)
       }
@@ -293,6 +297,7 @@ export default class ZoomPanSelection extends Toolbar {
       // change styles based on zoom or selection
       // zoom is Enabled and user has dragged, so draw blue rect
       if (w.globals.zoomEnabled && this.dragged) {
+        if (width < 0) width = 1 // fixes apexcharts.js#1168
         zoomRect.attr({
           x,
           y,
@@ -484,15 +489,10 @@ export default class ZoomPanSelection extends Toolbar {
 
     w.config.yaxis.forEach((yaxe, index) => {
       yHighestValue.push(
-        Math.floor(
-          w.globals.yAxisScale[index].niceMax -
-            xyRatios.yRatio[index] * me.startY
-        )
+        w.globals.yAxisScale[index].niceMax - xyRatios.yRatio[index] * me.startY
       )
       yLowestValue.push(
-        Math.floor(
-          w.globals.yAxisScale[index].niceMax - xyRatios.yRatio[index] * me.endY
-        )
+        w.globals.yAxisScale[index].niceMax - xyRatios.yRatio[index] * me.endY
       )
     })
 
@@ -511,6 +511,19 @@ export default class ZoomPanSelection extends Toolbar {
           w.globals.lastYAxis = Utils.clone(w.config.yaxis)
         }
 
+        if (w.config.xaxis.convertedCatToNumeric) {
+          xLowestValue = Math.floor(xLowestValue)
+          xHighestValue = Math.floor(xHighestValue)
+
+          if (xLowestValue < 1) {
+            xLowestValue = 1
+            xHighestValue = w.globals.dataPoints
+          }
+
+          if (xHighestValue - xLowestValue < 2) {
+            xHighestValue = xLowestValue + 1
+          }
+        }
         let xaxis = {
           min: xLowestValue,
           max: xHighestValue
@@ -547,7 +560,7 @@ export default class ZoomPanSelection extends Toolbar {
           // fix issue #650
           options.yaxis = yaxis
         }
-        me.ctx._updateOptions(
+        me.ctx.updateHelpers._updateOptions(
           options,
           false,
           me.w.config.chart.animations.dynamicAnimation.enabled
@@ -588,8 +601,6 @@ export default class ZoomPanSelection extends Toolbar {
     const w = this.w
     let me = context
 
-    let moveDirection
-
     // check to make sure there is data to compare against
     if (typeof w.globals.lastClientPosition.x !== 'undefined') {
       // get the change from last position to this position
@@ -598,13 +609,13 @@ export default class ZoomPanSelection extends Toolbar {
 
       // check which direction had the highest amplitude and then figure out direction by checking if the value is greater or less than zero
       if (Math.abs(deltaX) > Math.abs(deltaY) && deltaX > 0) {
-        moveDirection = 'left'
+        this.moveDirection = 'left'
       } else if (Math.abs(deltaX) > Math.abs(deltaY) && deltaX < 0) {
-        moveDirection = 'right'
+        this.moveDirection = 'right'
       } else if (Math.abs(deltaY) > Math.abs(deltaX) && deltaY > 0) {
-        moveDirection = 'up'
+        this.moveDirection = 'up'
       } else if (Math.abs(deltaY) > Math.abs(deltaX) && deltaY < 0) {
-        moveDirection = 'down'
+        this.moveDirection = 'down'
       }
     }
 
@@ -615,23 +626,51 @@ export default class ZoomPanSelection extends Toolbar {
     }
 
     let xLowestValue = w.globals.minX
+
     let xHighestValue = w.globals.maxX
 
-    me.panScrolled(moveDirection, xLowestValue, xHighestValue)
+    // on a category, we don't pan continuosly as it causes bugs
+    if (w.config.xaxis.convertedCatToNumeric) {
+      me.panScrolled(xLowestValue, xHighestValue)
+    }
   }
 
-  panScrolled(moveDirection, xLowestValue, xHighestValue) {
+  delayedPanScrolled() {
+    const w = this.w
+
+    let newMinX = w.globals.minX
+    let newMaxX = w.globals.maxX
+    const centerX = (w.globals.minX + w.globals.maxX) / 2
+
+    if (this.moveDirection === 'left') {
+      newMinX = w.globals.minX - centerX
+      newMaxX = w.globals.maxX - centerX
+    } else if (this.moveDirection === 'right') {
+      newMinX = w.globals.minX + centerX
+      newMaxX = w.globals.maxX + centerX
+    }
+
+    newMinX = Math.floor(newMinX)
+    newMaxX = Math.floor(newMaxX)
+    this.updateScrolledChart(
+      { xaxis: { min: newMinX, max: newMaxX } },
+      newMinX,
+      newMaxX
+    )
+  }
+
+  panScrolled(xLowestValue, xHighestValue) {
     const w = this.w
 
     const xyRatios = this.xyRatios
     let yaxis = Utils.clone(w.globals.initialConfig.yaxis)
 
-    if (moveDirection === 'left') {
+    if (this.moveDirection === 'left') {
       xLowestValue =
         w.globals.minX + (w.globals.gridWidth / 15) * xyRatios.xRatio
       xHighestValue =
         w.globals.maxX + (w.globals.gridWidth / 15) * xyRatios.xRatio
-    } else if (moveDirection === 'right') {
+    } else if (this.moveDirection === 'right') {
       xLowestValue =
         w.globals.minX - (w.globals.gridWidth / 15) * xyRatios.xRatio
       xHighestValue =
@@ -639,8 +678,8 @@ export default class ZoomPanSelection extends Toolbar {
     }
 
     if (
-      xLowestValue < w.globals.initialminX ||
-      xHighestValue > w.globals.initialmaxX
+      xLowestValue < w.globals.initialMinX ||
+      xHighestValue > w.globals.initialMaxX
     ) {
       xLowestValue = w.globals.minX
       xHighestValue = w.globals.maxX
@@ -670,7 +709,14 @@ export default class ZoomPanSelection extends Toolbar {
       // fix issue #650
       options.yaxis = yaxis
     }
-    this.ctx._updateOptions(options, false, false)
+
+    this.updateScrolledChart(options, xLowestValue, xHighestValue)
+  }
+
+  updateScrolledChart(options, xLowestValue, xHighestValue) {
+    const w = this.w
+
+    this.ctx.updateHelpers._updateOptions(options, false, false)
 
     if (typeof w.config.chart.events.scrolled === 'function') {
       w.config.chart.events.scrolled(this.ctx, {
