@@ -8,10 +8,10 @@ export default class AxesUtils {
   }
 
   // Based on the formatter function, get the label text and position
-  getLabel(labels, timelineLabels, x, i, drawnLabels = []) {
+  getLabel(labels, timescaleLabels, x, i, drawnLabels = [], fontSize = '12px') {
     const w = this.w
     let rawLabel = typeof labels[i] === 'undefined' ? '' : labels[i]
-    let label
+    let label = rawLabel
 
     let xlbFormatter = w.globals.xLabelFormatter
     let customFormatter = w.config.xaxis.labels.formatter
@@ -28,7 +28,7 @@ export default class AxesUtils {
 
     const determineHighestUnit = (unit) => {
       let highestUnit = null
-      timelineLabels.forEach((t) => {
+      timescaleLabels.forEach((t) => {
         if (t.unit === 'month') {
           highestUnit = 'year'
         } else if (t.unit === 'day') {
@@ -42,10 +42,10 @@ export default class AxesUtils {
 
       return highestUnit === unit
     }
-    if (timelineLabels.length > 0) {
-      isBold = determineHighestUnit(timelineLabels[i].unit)
-      x = timelineLabels[i].position
-      label = timelineLabels[i].value
+    if (timescaleLabels.length > 0) {
+      isBold = determineHighestUnit(timescaleLabels[i].unit)
+      x = timescaleLabels[i].position
+      label = timescaleLabels[i].value
     } else {
       if (w.config.xaxis.type === 'datetime' && customFormatter === undefined) {
         label = ''
@@ -54,13 +54,29 @@ export default class AxesUtils {
 
     if (typeof label === 'undefined') label = ''
 
-    label = label.toString()
+    label = Array.isArray(label) ? label : label.toString()
+
+    let graphics = new Graphics(this.ctx)
+    let textRect = {}
+    if (w.globals.rotateXLabels) {
+      textRect = graphics.getTextRects(
+        label,
+        parseInt(fontSize, 10),
+        null,
+        `rotate(${w.config.xaxis.labels.rotate} 0 0)`,
+        false
+      )
+    } else {
+      textRect = graphics.getTextRects(label, parseInt(fontSize, 10))
+    }
 
     if (
-      label.indexOf('NaN') === 0 ||
-      label.toLowerCase().indexOf('invalid') === 0 ||
-      label.toLowerCase().indexOf('infinity') >= 0 ||
-      (drawnLabels.indexOf(label) >= 0 && !w.config.xaxis.labels.showDuplicates)
+      !Array.isArray(label) &&
+      (label.indexOf('NaN') === 0 ||
+        label.toLowerCase().indexOf('invalid') === 0 ||
+        label.toLowerCase().indexOf('infinity') >= 0 ||
+        (drawnLabels.indexOf(label) >= 0 &&
+          !w.config.xaxis.labels.showDuplicates))
     ) {
       label = ''
     }
@@ -68,8 +84,57 @@ export default class AxesUtils {
     return {
       x,
       text: label,
+      textRect,
       isBold
     }
+  }
+
+  checkForOverflowingLabels(
+    i,
+    label,
+    labelsLen,
+    drawnLabels,
+    drawnLabelsRects
+  ) {
+    const w = this.w
+
+    if (i === 0) {
+      // check if first label is being truncated
+      if (w.globals.skipFirstTimelinelabel) {
+        label.text = ''
+      }
+    }
+
+    if (i === labelsLen - 1) {
+      // check if last label is being truncated
+      if (w.globals.skipLastTimelinelabel) {
+        label.text = ''
+      }
+    }
+
+    if (w.config.xaxis.labels.hideOverlappingLabels && drawnLabels.length > 0) {
+      const prev = drawnLabelsRects[drawnLabelsRects.length - 1]
+      if (
+        label.x <
+        prev.textRect.width /
+          (w.globals.rotateXLabels
+            ? Math.abs(w.config.xaxis.labels.rotate) / 12
+            : 1.1) +
+          prev.x
+      ) {
+        label.text = ''
+      }
+    }
+
+    return label
+  }
+
+  checkForReversedLabels(i, labels) {
+    const w = this.w
+    if (w.config.yaxis[i] && w.config.yaxis[i].reversed) {
+      labels.reverse()
+    }
+    return labels
   }
 
   drawYAxisTicks(
@@ -87,7 +152,7 @@ export default class AxesUtils {
     // initial label position = 0;
     let t = w.globals.translateY
 
-    if (axisTicks.show) {
+    if (axisTicks.show && tickAmount > 0) {
       if (w.config.yaxis[realIndex].opposite === true) x = x + axisTicks.width
 
       for (let i = tickAmount; i >= 0; i--) {
@@ -96,12 +161,16 @@ export default class AxesUtils {
         if (w.globals.isBarHorizontal) {
           tY = labelsDivider * i
         }
+
+        if (w.config.chart.type === 'heatmap') {
+          tY = tY + labelsDivider / 2
+        }
         let elTick = graphics.drawLine(
           x + axisBorder.offsetX - axisTicks.width + axisTicks.offsetX,
           tY + axisTicks.offsetY,
           x + axisBorder.offsetX + axisTicks.offsetX,
           tY + axisTicks.offsetY,
-          axisBorder.color
+          axisTicks.color
         )
         elYaxis.add(elTick)
         t = t + labelsDivider
