@@ -13,13 +13,16 @@ export default class XAxis {
     this.w = ctx.w
 
     const w = this.w
+    this.axesUtils = new AxesUtils(ctx)
+
     this.xaxisLabels = w.globals.labels.slice()
-    if (w.globals.timelineLabels.length > 0) {
-      //  timeline labels are there
-      this.xaxisLabels = w.globals.timelineLabels.slice()
+    if (w.globals.timescaleLabels.length > 0 && !w.globals.isBarHorizontal) {
+      //  timeline labels are there and chart is not rangeabr timeline
+      this.xaxisLabels = w.globals.timescaleLabels.slice()
     }
 
     this.drawnLabels = []
+    this.drawnLabelsRects = []
 
     if (w.config.xaxis.position === 'top') {
       this.offY = 0
@@ -27,24 +30,28 @@ export default class XAxis {
       this.offY = w.globals.gridHeight + 1
     }
     this.offY = this.offY + w.config.xaxis.axisBorder.offsetY
+    this.isCategoryBarHorizontal =
+      w.config.chart.type === 'bar' && w.config.plotOptions.bar.horizontal
 
     this.xaxisFontSize = w.config.xaxis.labels.style.fontSize
     this.xaxisFontFamily = w.config.xaxis.labels.style.fontFamily
     this.xaxisForeColors = w.config.xaxis.labels.style.colors
     this.xaxisBorderWidth = w.config.xaxis.axisBorder.width
+    if (this.isCategoryBarHorizontal) {
+      this.xaxisBorderWidth = w.config.yaxis[0].axisBorder.width.toString()
+    }
 
     if (this.xaxisBorderWidth.indexOf('%') > -1) {
       this.xaxisBorderWidth =
-        (w.globals.gridWidth * parseInt(this.xaxisBorderWidth)) / 100
+        (w.globals.gridWidth * parseInt(this.xaxisBorderWidth, 10)) / 100
     } else {
-      this.xaxisBorderWidth = parseInt(this.xaxisBorderWidth)
+      this.xaxisBorderWidth = parseInt(this.xaxisBorderWidth, 10)
     }
     this.xaxisBorderHeight = w.config.xaxis.axisBorder.height
 
     // For bars, we will only consider single y xais,
     // as we are not providing multiple yaxis for bar charts
     this.yaxis = w.config.yaxis[0]
-    this.axesUtils = new AxesUtils(ctx)
   }
 
   drawXaxis() {
@@ -53,16 +60,12 @@ export default class XAxis {
 
     let elXaxis = graphics.group({
       class: 'apexcharts-xaxis',
-      transform: `translate(${w.config.xaxis.offsetX}, ${
-        w.config.xaxis.offsetY
-      })`
+      transform: `translate(${w.config.xaxis.offsetX}, ${w.config.xaxis.offsetY})`
     })
 
     let elXaxisTexts = graphics.group({
       class: 'apexcharts-xaxis-texts-g',
-      transform: `translate(${w.globals.translateXAxisX}, ${
-        w.globals.translateXAxisY
-      })`
+      transform: `translate(${w.globals.translateXAxisX}, ${w.globals.translateXAxisY})`
     })
 
     elXaxis.add(elXaxisTexts)
@@ -78,7 +81,8 @@ export default class XAxis {
     }
 
     if (w.globals.isXNumeric) {
-      colWidth = w.globals.gridWidth / (labels.length - 1)
+      let len = labels.length > 1 ? labels.length - 1 : labels.length
+      colWidth = w.globals.gridWidth / len
       xPos = xPos + colWidth / 2 + w.config.xaxis.labels.offsetX
     } else {
       colWidth = w.globals.gridWidth / labels.length
@@ -91,49 +95,68 @@ export default class XAxis {
       for (let i = 0; i <= labelsLen - 1; i++) {
         let x = xPos - colWidth / 2 + w.config.xaxis.labels.offsetX
 
+        if (
+          i === 0 &&
+          labelsLen === 1 &&
+          colWidth / 2 === xPos &&
+          w.globals.dataPoints === 1
+        ) {
+          // single datapoint
+          x = w.globals.gridWidth / 2
+        }
         let label = this.axesUtils.getLabel(
           labels,
-          w.globals.timelineLabels,
+          w.globals.timescaleLabels,
           x,
           i,
-          this.drawnLabels
+          this.drawnLabels,
+          this.xaxisFontSize
         )
-
-        this.drawnLabels.push(label.text)
 
         let offsetYCorrection = 28
         if (w.globals.rotateXLabels) {
           offsetYCorrection = 22
         }
-        let elTick = graphics.drawText({
+
+        label = this.axesUtils.checkForOverflowingLabels(
+          i,
+          label,
+          labelsLen,
+          this.drawnLabels,
+          this.drawnLabelsRects
+        )
+
+        const getCatForeColor = () => {
+          return w.config.xaxis.convertedCatToNumeric
+            ? this.xaxisForeColors[w.globals.minX + i - 1]
+            : this.xaxisForeColors[i]
+        }
+
+        let elText = graphics.drawText({
           x: label.x,
           y: this.offY + w.config.xaxis.labels.offsetY + offsetYCorrection,
-          text: '',
+          text: label.text,
           textAnchor: 'middle',
           fontWeight: label.isBold ? 600 : 400,
           fontSize: this.xaxisFontSize,
           fontFamily: this.xaxisFontFamily,
           foreColor: Array.isArray(this.xaxisForeColors)
-            ? this.xaxisForeColors[i]
+            ? getCatForeColor()
             : this.xaxisForeColors,
+          isPlainText: false,
           cssClass:
             'apexcharts-xaxis-label ' + w.config.xaxis.labels.style.cssClass
         })
 
-        if (i === labelsLen - 1) {
-          if (w.globals.skipLastTimelinelabel) {
-            label.text = ''
-          }
-        }
-
-        elXaxisTexts.add(elTick)
-
-        graphics.addTspan(elTick, label.text, this.xaxisFontFamily)
+        elXaxisTexts.add(elText)
 
         let elTooltipTitle = document.createElementNS(w.globals.SVGNS, 'title')
         elTooltipTitle.textContent = label.text
-        elTick.node.appendChild(elTooltipTitle)
-
+        elText.node.appendChild(elTooltipTitle)
+        if (label.text !== '') {
+          this.drawnLabels.push(label.text)
+          this.drawnLabelsRects.push(label)
+        }
         xPos = xPos + colWidth
       }
     }
@@ -147,7 +170,7 @@ export default class XAxis {
         x: w.globals.gridWidth / 2 + w.config.xaxis.title.offsetX,
         y:
           this.offY -
-          parseInt(this.xaxisFontSize) +
+          parseFloat(this.xaxisFontSize) +
           w.globals.xAxisLabelsHeight +
           w.config.xaxis.title.offsetY,
         text: w.config.xaxis.title.text,
@@ -214,8 +237,10 @@ export default class XAxis {
     let yPos
     let labels = []
 
-    for (let i = 0; i < this.xaxisLabels.length; i++) {
-      labels.push(this.xaxisLabels[i])
+    if (w.config.yaxis[realIndex].show) {
+      for (let i = 0; i < this.xaxisLabels.length; i++) {
+        labels.push(this.xaxisLabels[i])
+      }
     }
 
     colHeight = w.globals.gridHeight / labels.length
@@ -229,11 +254,19 @@ export default class XAxis {
       for (let i = 0; i <= labels.length - 1; i++) {
         let label = typeof labels[i] === 'undefined' ? '' : labels[i]
 
-        label = lbFormatter(label)
+        label = lbFormatter(label, {
+          seriesIndex: realIndex,
+          dataPointIndex: i,
+          w
+        })
 
+        let multiY = 0
+        if (Array.isArray(label)) {
+          multiY = (label.length / 2) * parseInt(ylabels.style.fontSize, 10)
+        }
         let elLabel = graphics.drawText({
           x: ylabels.offsetX - 15,
-          y: yPos + colHeight + ylabels.offsetY,
+          y: yPos + colHeight + ylabels.offsetY - multiY,
           text: label,
           textAnchor: this.yaxis.opposite ? 'start' : 'end',
           foreColor: ylabels.style.color
@@ -241,18 +274,21 @@ export default class XAxis {
             : ylabels.style.colors[i],
           fontSize: ylabels.style.fontSize,
           fontFamily: ylabels.style.fontFamily,
+          isPlainText: false,
           cssClass: 'apexcharts-yaxis-label ' + ylabels.style.cssClass
         })
 
         elYaxisTexts.add(elLabel)
 
+        let elTooltipTitle = document.createElementNS(w.globals.SVGNS, 'title')
+        elTooltipTitle.textContent = label.text
+        elLabel.node.appendChild(elTooltipTitle)
+
         if (w.config.yaxis[realIndex].labels.rotate !== 0) {
           let labelRotatingCenter = graphics.rotateAroundCenter(elLabel.node)
           elLabel.node.setAttribute(
             'transform',
-            `rotate(${w.config.yaxis[realIndex].labels.rotate} ${
-              labelRotatingCenter.x
-            } ${labelRotatingCenter.y})`
+            `rotate(${w.config.yaxis[realIndex].labels.rotate} 0 ${labelRotatingCenter.y})`
           )
         }
         yPos = yPos + colHeight
@@ -283,21 +319,27 @@ export default class XAxis {
       elYaxis.add(elXaxisTitle)
     }
 
-    if (w.config.xaxis.axisBorder.show) {
-      let elHorzLine = graphics.drawLine(
-        w.globals.padHorizontal + w.config.xaxis.axisBorder.offsetX,
-        this.offY,
-        this.xaxisBorderWidth,
-        this.offY,
-        this.yaxis.axisBorder.color,
-        0,
-        this.xaxisBorderHeight
+    let offX = 0
+    if (this.isCategoryBarHorizontal && w.config.yaxis[0].opposite) {
+      offX = w.globals.gridWidth
+    }
+    const axisBorder = w.config.xaxis.axisBorder
+    if (axisBorder.show) {
+      let elVerticalLine = graphics.drawLine(
+        w.globals.padHorizontal + axisBorder.offsetX + offX,
+        1 + axisBorder.offsetY,
+        w.globals.padHorizontal + axisBorder.offsetX + offX,
+        w.globals.gridHeight + axisBorder.offsetY,
+        axisBorder.color,
+        0
       )
 
-      elYaxis.add(elHorzLine)
+      elYaxis.add(elVerticalLine)
+    }
 
+    if (w.config.yaxis[0].axisTicks.show) {
       this.axesUtils.drawYAxisTicks(
-        0,
+        offX,
         labels.length,
         w.config.yaxis[0].axisBorder,
         w.config.yaxis[0].axisTicks,
@@ -314,7 +356,7 @@ export default class XAxis {
     let w = this.w
     let x2 = x1
 
-    if (x1 < 0 || x1 > w.globals.gridWidth) return
+    if (x1 < 0 || x1 - 2 > w.globals.gridWidth) return
 
     let y1 = this.offY + w.config.xaxis.axisTicks.offsetY
     let y2 = y1 + w.config.xaxis.axisTicks.height
@@ -343,7 +385,7 @@ export default class XAxis {
     const xCount = this.xaxisLabels.length
     let x1 = w.globals.padHorizontal
 
-    if (w.globals.timelineLabels.length > 0) {
+    if (w.globals.timescaleLabels.length > 0) {
       for (let i = 0; i < xCount; i++) {
         x1 = this.xaxisLabels[i].position
         xAxisTicksPositions.push(x1)
@@ -378,7 +420,7 @@ export default class XAxis {
       '.apexcharts-yaxis-inversed text'
     )
     let xAxisTextsInversed = w.globals.dom.baseEl.querySelectorAll(
-      '.apexcharts-xaxis-inversed-texts-g text'
+      '.apexcharts-xaxis-inversed-texts-g text tspan'
     )
 
     if (w.globals.rotateXLabels || w.config.xaxis.labels.rotateAlways) {
@@ -389,9 +431,7 @@ export default class XAxis {
 
         xAxisTexts[xat].setAttribute(
           'transform',
-          `rotate(${w.config.xaxis.labels.rotate} ${textRotatingCenter.x} ${
-            textRotatingCenter.y
-          })`
+          `rotate(${w.config.xaxis.labels.rotate} ${textRotatingCenter.x} ${textRotatingCenter.y})`
         )
 
         xAxisTexts[xat].setAttribute('text-anchor', `end`)
@@ -403,21 +443,26 @@ export default class XAxis {
         let tSpan = xAxisTexts[xat].childNodes
 
         if (w.config.xaxis.labels.trim) {
-          graphics.placeTextWithEllipsis(
-            tSpan[0],
-            tSpan[0].textContent,
-            w.config.xaxis.labels.maxHeight - 40
-          )
+          Array.prototype.forEach.call(tSpan, (ts) => {
+            graphics.placeTextWithEllipsis(
+              ts,
+              ts.textContent,
+              w.config.xaxis.labels.maxHeight -
+                (w.config.legend.position === 'bottom' ? 20 : 10)
+            )
+          })
         }
       }
     } else {
-      let width = w.globals.gridWidth / w.globals.labels.length
+      let width = w.globals.gridWidth / (w.globals.labels.length + 1)
 
       for (let xat = 0; xat < xAxisTexts.length; xat++) {
         let tSpan = xAxisTexts[xat].childNodes
 
         if (w.config.xaxis.labels.trim && w.config.xaxis.type !== 'datetime') {
-          graphics.placeTextWithEllipsis(tSpan[0], tSpan[0].textContent, width)
+          Array.prototype.forEach.call(tSpan, (ts) => {
+            graphics.placeTextWithEllipsis(ts, ts.textContent, width)
+          })
         }
       }
     }
@@ -437,7 +482,10 @@ export default class XAxis {
         )
       }
 
-      if (lastLabelPosX.x + lastLabelPosX.width > w.globals.gridWidth) {
+      if (
+        lastLabelPosX.x + lastLabelPosX.width > w.globals.gridWidth &&
+        !w.globals.isBarHorizontal
+      ) {
         yAxisTextsInversed[0].parentNode.removeChild(yAxisTextsInversed[0])
       }
 
@@ -447,7 +495,7 @@ export default class XAxis {
           xAxisTextsInversed[xat],
           xAxisTextsInversed[xat].textContent,
           w.config.yaxis[0].labels.maxWidth -
-            parseInt(w.config.yaxis[0].title.style.fontSize) * 2 -
+            parseFloat(w.config.yaxis[0].title.style.fontSize) * 2 -
             20
         )
       }
