@@ -18,6 +18,8 @@ export default class YAxis {
     this.axisFontFamily = w.config.xaxis.labels.style.fontFamily
 
     this.xaxisForeColors = w.config.xaxis.labels.style.colors
+    this.isCategoryBarHorizontal =
+      w.config.chart.type === 'bar' && w.config.plotOptions.bar.horizontal
 
     this.xAxisoffX = 0
     if (w.config.xaxis.position === 'bottom') {
@@ -53,17 +55,17 @@ export default class YAxis {
     let tickAmount = w.globals.yAxisScale[realIndex].result.length - 1
 
     // labelsDivider is simply svg height/number of ticks
-    let labelsDivider = w.globals.gridHeight / tickAmount + 0.1
+    let labelsDivider = w.globals.gridHeight / tickAmount
 
     // initial label position = 0;
     let l = w.globals.translateY
     let lbFormatter = w.globals.yLabelFormatters[realIndex]
 
     let labels = w.globals.yAxisScale[realIndex].result.slice()
-    if (w.config.yaxis[realIndex] && w.config.yaxis[realIndex].reversed) {
-      labels.reverse()
-    }
 
+    labels = this.axesUtils.checkForReversedLabels(realIndex, labels)
+
+    let firstLabel = ''
     if (w.config.yaxis[realIndex].labels.show) {
       for (let i = tickAmount; i >= 0; i--) {
         let val = labels[i]
@@ -83,19 +85,24 @@ export default class YAxis {
           fontSize: yaxisFontSize,
           fontFamily: yaxisFontFamily,
           foreColor: w.config.yaxis[realIndex].labels.style.color,
+          isPlainText: false,
           cssClass:
             'apexcharts-yaxis-label ' +
             w.config.yaxis[realIndex].labels.style.cssClass
         })
+        if (i === tickAmount) {
+          firstLabel = label
+        }
         elYaxisTexts.add(label)
 
-        let labelRotatingCenter = graphics.rotateAroundCenter(label.node)
         if (w.config.yaxis[realIndex].labels.rotate !== 0) {
+          let firstabelRotatingCenter = graphics.rotateAroundCenter(
+            firstLabel.node
+          )
+          let labelRotatingCenter = graphics.rotateAroundCenter(label.node)
           label.node.setAttribute(
             'transform',
-            `rotate(${w.config.yaxis[realIndex].labels.rotate} ${
-              labelRotatingCenter.x
-            } ${labelRotatingCenter.y})`
+            `rotate(${w.config.yaxis[realIndex].labels.rotate} ${firstabelRotatingCenter.x} ${labelRotatingCenter.y})`
           )
         }
         l = l + labelsDivider
@@ -113,7 +120,10 @@ export default class YAxis {
       }
       let elYAxisTitleText = graphics.drawText({
         x,
-        y: w.globals.gridHeight / 2 + w.globals.translateY,
+        y:
+          w.globals.gridHeight / 2 +
+          w.globals.translateY +
+          w.config.yaxis[realIndex].title.offsetY,
         text: w.config.yaxis[realIndex].title.text,
         textAnchor: 'end',
         foreColor: w.config.yaxis[realIndex].title.style.color,
@@ -130,22 +140,26 @@ export default class YAxis {
     }
 
     let axisBorder = w.config.yaxis[realIndex].axisBorder
-    if (axisBorder.show) {
-      let x = 31 + axisBorder.offsetX
-      if (w.config.yaxis[realIndex].opposite) {
-        x = -31 - axisBorder.offsetX
-      }
 
+    let x = 31 + axisBorder.offsetX
+    if (w.config.yaxis[realIndex].opposite) {
+      x = -31 - axisBorder.offsetX
+    }
+
+    if (axisBorder.show) {
       let elVerticalLine = graphics.drawLine(
         x,
         w.globals.translateY + axisBorder.offsetY - 2,
         x,
         w.globals.gridHeight + w.globals.translateY + axisBorder.offsetY + 2,
-        axisBorder.color
+        axisBorder.color,
+        0,
+        axisBorder.width
       )
 
       elYaxis.add(elVerticalLine)
-
+    }
+    if (w.config.yaxis[realIndex].axisTicks.show) {
       this.axesUtils.drawYAxisTicks(
         x,
         tickAmount,
@@ -171,9 +185,7 @@ export default class YAxis {
 
     let elXaxisTexts = graphics.group({
       class: 'apexcharts-xaxis-texts-g',
-      transform: `translate(${w.globals.translateXAxisX}, ${
-        w.globals.translateXAxisY
-      })`
+      transform: `translate(${w.globals.translateXAxisX}, ${w.globals.translateXAxisY})`
     })
 
     elXaxis.add(elXaxisTexts)
@@ -190,25 +202,19 @@ export default class YAxis {
 
     let labels = w.globals.yAxisScale[realIndex].result.slice()
 
-    let timelineLabels = w.globals.invertedTimelineLabels
-    if (timelineLabels.length > 0) {
-      this.xaxisLabels = timelineLabels.slice()
-      labels = timelineLabels.slice()
+    let timescaleLabels = w.globals.timescaleLabels
+    if (timescaleLabels.length > 0) {
+      this.xaxisLabels = timescaleLabels.slice()
+      labels = timescaleLabels.slice()
       tickAmount = labels.length
     }
 
-    if (w.config.yaxis[realIndex] && w.config.yaxis[realIndex].reversed) {
-      labels.reverse()
-    }
+    labels = this.axesUtils.checkForReversedLabels(realIndex, labels)
 
-    const tl = timelineLabels.length
+    const tl = timescaleLabels.length
 
     if (w.config.xaxis.labels.show) {
-      for (
-        let i = tl ? 0 : tickAmount;
-        tl ? i < tl - 1 : i >= 0;
-        tl ? i++ : i--
-      ) {
+      for (let i = tl ? 0 : tickAmount; tl ? i < tl : i >= 0; tl ? i++ : i--) {
         let val = labels[i]
         val = lbFormatter(val, i)
 
@@ -217,29 +223,37 @@ export default class YAxis {
           w.globals.padHorizontal -
           (l - labelsDivider + w.config.xaxis.labels.offsetX)
 
-        if (timelineLabels.length) {
+        if (timescaleLabels.length) {
           let label = this.axesUtils.getLabel(
             labels,
-            timelineLabels,
+            timescaleLabels,
             x,
             i,
-            this.drawnLabels
+            this.drawnLabels,
+            this.xaxisFontSize
           )
           x = label.x
           val = label.text
           this.drawnLabels.push(label.text)
-        }
 
+          if (i === 0 && w.globals.skipFirstTimelinelabel) {
+            val = ''
+          }
+          if (i === labels.length - 1 && w.globals.skipLastTimelinelabel) {
+            val = ''
+          }
+        }
         let elTick = graphics.drawText({
-          x: x,
+          x,
           y: this.xAxisoffX + w.config.xaxis.labels.offsetY + 30,
-          text: '',
+          text: val,
           textAnchor: 'middle',
           foreColor: Array.isArray(this.xaxisForeColors)
             ? this.xaxisForeColors[realIndex]
             : this.xaxisForeColors,
           fontSize: this.xaxisFontSize,
           fontFamily: this.xaxisFontFamily,
+          isPlainText: false,
           cssClass:
             'apexcharts-xaxis-label ' + w.config.xaxis.labels.style.cssClass
         })
@@ -256,6 +270,39 @@ export default class YAxis {
       }
     }
 
+    this.inversedYAxisTitleText(elXaxis)
+    this.inversedYAxisBorder(elXaxis)
+
+    return elXaxis
+  }
+
+  inversedYAxisBorder(parent) {
+    const w = this.w
+    const graphics = new Graphics(this.ctx)
+
+    let axisBorder = w.config.xaxis.axisBorder
+    if (axisBorder.show) {
+      let lineCorrection = 0
+      if (w.config.chart.type === 'bar' && w.globals.isXNumeric) {
+        lineCorrection = lineCorrection - 15
+      }
+      let elHorzLine = graphics.drawLine(
+        w.globals.padHorizontal + lineCorrection + axisBorder.offsetX,
+        this.xAxisoffX,
+        w.globals.gridWidth,
+        this.xAxisoffX,
+        axisBorder.color,
+        0,
+        axisBorder.height
+      )
+
+      parent.add(elHorzLine)
+    }
+  }
+
+  inversedYAxisTitleText(parent) {
+    const w = this.w
+    const graphics = new Graphics(this.ctx)
     if (w.config.xaxis.title.text !== undefined) {
       let elYaxisTitle = graphics.group({
         class: 'apexcharts-xaxis-title apexcharts-yaxis-title-inversed'
@@ -265,8 +312,8 @@ export default class YAxis {
         x: w.globals.gridWidth / 2,
         y:
           this.xAxisoffX +
-          parseInt(this.xaxisFontSize) +
-          parseInt(w.config.xaxis.title.style.fontSize) +
+          parseFloat(this.xaxisFontSize) +
+          parseFloat(w.config.xaxis.title.style.fontSize) +
           20,
         text: w.config.xaxis.title.text,
         textAnchor: 'middle',
@@ -278,23 +325,8 @@ export default class YAxis {
 
       elYaxisTitle.add(elYAxisTitleText)
 
-      elXaxis.add(elYaxisTitle)
+      parent.add(elYaxisTitle)
     }
-
-    let axisBorder = w.config.yaxis[realIndex].axisBorder
-    if (axisBorder.show) {
-      let elVerticalLine = graphics.drawLine(
-        w.globals.padHorizontal + axisBorder.offsetX,
-        1 + axisBorder.offsetY,
-        w.globals.padHorizontal + axisBorder.offsetX,
-        w.globals.gridHeight + axisBorder.offsetY,
-        axisBorder.color
-      )
-
-      elXaxis.add(elVerticalLine)
-    }
-
-    return elXaxis
   }
 
   yAxisTitleRotate(realIndex, yAxisOpposite) {
@@ -340,21 +372,12 @@ export default class YAxis {
 
     if (yAxisTitle !== null) {
       let titleRotatingCenter = graphics.rotateAroundCenter(yAxisTitle)
-      if (!yAxisOpposite) {
-        yAxisTitle.setAttribute(
-          'transform',
-          `rotate(-${w.config.yaxis[realIndex].title.rotate} ${
-            titleRotatingCenter.x
-          } ${titleRotatingCenter.y})`
-        )
-      } else {
-        yAxisTitle.setAttribute(
-          'transform',
-          `rotate(${w.config.yaxis[realIndex].title.rotate} ${
-            titleRotatingCenter.x
-          } ${titleRotatingCenter.y})`
-        )
-      }
+      yAxisTitle.setAttribute(
+        'transform',
+        `rotate(${yAxisOpposite ? '' : '-'}${
+          w.config.yaxis[realIndex].title.rotate
+        } ${titleRotatingCenter.x} ${titleRotatingCenter.y})`
+      )
     }
   }
 
@@ -405,7 +428,10 @@ export default class YAxis {
       }
     }
 
-    return { xPos: x, padd }
+    return {
+      xPos: x,
+      padd
+    }
   }
 
   // sets the x position of the y-axis by counting the labels width, title width and any offset
@@ -414,7 +440,7 @@ export default class YAxis {
 
     let xLeft = 0
     let xRight = 0
-    let leftOffsetX = 21
+    let leftOffsetX = 18
     let rightOffsetX = 1
 
     if (w.config.yaxis.length > 1) {
