@@ -1,6 +1,70 @@
 import Utils from './../../utils/Utils'
 
 export default class Globals {
+  initGlobalVars(gl) {
+    gl.series = [] // the MAIN series array (y values)
+    gl.seriesCandleO = []
+    gl.seriesCandleH = []
+    gl.seriesCandleL = []
+    gl.seriesCandleC = []
+    gl.seriesRangeStart = []
+    gl.seriesRangeEnd = []
+    gl.seriesRangeBarTimeline = []
+    gl.seriesPercent = []
+    gl.seriesX = []
+    gl.seriesZ = []
+    gl.seriesNames = []
+    gl.seriesTotals = []
+    gl.seriesLog = []
+    gl.stackedSeriesTotals = []
+    gl.seriesXvalues = [] // we will need this in tooltip (it's x position)
+    // when we will have unequal x values, we will need
+    // some way to get x value depending on mouse pointer
+    gl.seriesYvalues = [] // we will need this when deciding which series
+    // user hovered on
+    gl.labels = []
+    gl.categoryLabels = []
+    gl.timescaleLabels = []
+    gl.noLabelsProvided = false
+    gl.resizeTimer = null
+    gl.selectionResizeTimer = null
+    gl.delayedElements = []
+    gl.pointsArray = []
+    gl.dataLabelsRects = []
+    gl.isXNumeric = false
+    gl.skipLastTimelinelabel = false
+    gl.skipFirstTimelinelabel = false
+    gl.x2SpaceAvailable = 0
+    gl.isDataXYZ = false
+    gl.isMultiLineX = false
+    gl.isMultipleYAxis = false
+    gl.maxY = -Number.MAX_VALUE
+    gl.minY = Number.MIN_VALUE
+    gl.minYArr = []
+    gl.maxYArr = []
+    gl.maxX = -Number.MAX_VALUE
+    gl.minX = Number.MAX_VALUE
+    gl.initialMaxX = -Number.MAX_VALUE
+    gl.initialMinX = Number.MAX_VALUE
+    gl.maxDate = 0
+    gl.minDate = Number.MAX_VALUE
+    gl.minZ = Number.MAX_VALUE
+    gl.maxZ = -Number.MAX_VALUE
+    gl.minXDiff = Number.MAX_VALUE
+    gl.yAxisScale = []
+    gl.xAxisScale = null
+    gl.xAxisTicksPositions = []
+    gl.yLabelsCoords = []
+    gl.yTitleCoords = []
+    gl.barPadForNumericAxis = 0
+    gl.padHorizontal = 0
+    gl.xRange = 0
+    gl.yRange = []
+    gl.zRange = 0
+    gl.dataPoints = 0
+    gl.xTickAmount = 0
+  }
+
   globalVars(config) {
     return {
       chartID: null, // chart ID - apexcharts-cuid
@@ -46,19 +110,10 @@ export default class Globals {
       initialConfig: null, // we will store the first config user has set to go back when user finishes interactions like zooming and come out of it
       lastXAxis: [],
       lastYAxis: [],
-      series: [], // the MAIN series array (y values)
-      seriesRangeStart: [], // the clone of series becomes the start in range
-      seriesRangeEnd: [], // the end values in range chart
-      seriesPercent: [], // the percentage values of the given series
-      seriesTotals: [],
-      stackedSeriesTotals: [],
-      seriesX: [], // store the numeric x values in this array (x values)
-      seriesZ: [], // The 3rd "Z" dimension for bubbles chart (z values)
+      columnSeries: null,
       labels: [], // store the text to draw on x axis
       // Don't mutate the labels, many things including tooltips depends on it!
-      timelineLabels: [], // store the timeline Labels in another variable
-      invertedTimelineLabels: [], // for rangebar timeline
-      seriesNames: [], // same as labels, used in non axis charts
+      timescaleLabels: [], // store the timescaleLabels Labels in another variable
       noLabelsProvided: false, // if user didn't provide any categories/labels or x values, fallback to 1,2,3,4...
       allSeriesCollapsed: false,
       collapsedSeries: [], // when user collapses a series, it goes into this array
@@ -70,9 +125,10 @@ export default class Globals {
       capturedSeriesIndex: -1,
       capturedDataPointIndex: -1,
       selectedDataPoints: [],
+      goldenPadding: 35, // this value is used at a lot of places for spacing purpose
+      invalidLogScale: false, // if a user enabled log scale but the data provided is not valid to generate a log scale, turn on this flag
       ignoreYAxisIndexes: [], // when series are being collapsed in multiple y axes, ignore certain index
       yAxisSameScaleIndices: [],
-      padHorizontal: 0,
       maxValsInArrayIndex: 0,
       radialSize: 0,
       zoomEnabled:
@@ -86,26 +142,9 @@ export default class Globals {
         config.chart.toolbar.autoSelected === 'selection' &&
         config.chart.toolbar.tools.selection,
       yaxis: null,
-      minY: Number.MIN_VALUE, //  is 5e-324, i.e. the smallest positive number
-      // NOTE: If there are multiple y axis, the first yaxis array element will be considered for all y values calculations. Rest all will be calculated based on that
-      maxY: -Number.MAX_VALUE, // is -1.7976931348623157e+308
-      // NOTE: The above note for minY applies here as well
-
-      minYArr: [],
-      maxYArr: [],
-      maxX: -Number.MAX_VALUE, // is -1.7976931348623157e+308
-      initialmaxX: -Number.MAX_VALUE,
-      minX: Number.MIN_VALUE, //  is 5e-324, i.e. the smallest positive number
-      initialminX: Number.MIN_VALUE,
-      minZ: Number.MIN_VALUE, // Max Z value in charts with Z axis
-      maxZ: -Number.MAX_VALUE, // Max Z value in charts with Z axis
-      minXDiff: Number.MAX_VALUE,
       mousedown: false,
       lastClientPosition: {}, // don't reset this variable this the chart is destroyed. It is used to detect right or left mousemove in panning
       visibleXRange: undefined,
-      yRange: [], // this property is the absolute sum of positive and negative values [eg (-100 + 200 = 300)] - yAxis
-      zRange: 0, // zAxis Range (for bubble charts)
-      xRange: 0, // xAxis range
       yValueDecimal: 0, // are there floating numbers in the series. If yes, this represent the len of the decimals
       total: 0,
       SVGNS: 'http://www.w3.org/2000/svg', // svg namespace
@@ -114,55 +153,34 @@ export default class Globals {
       noData: false, // whether there is any data to display or not
       locale: {}, // the current locale values will be preserved here for global access
       dom: {}, // for storing all dom nodes in this particular property
-      // elWrap: null, // the element that wraps everything
-      // elGraphical: null, // this contains lines/areas/bars/pies
-      // elGridRect: null, // paths going outside this area will be clipped
-      // elGridRectMask: null, // clipping will happen with this mask
-      // elGridRectMarkerMask: null, // clipping will happen with this mask
-      // elLegendWrap: null, // the whole legend area
-      // elDefs: null, // [defs] element
       memory: {
         methodsToExec: []
       },
       shouldAnimate: true,
       skipLastTimelinelabel: false, // when last label is cropped, skip drawing it
+      skipFirstTimelinelabel: false, // when first label is cropped, skip drawing it
       delayedElements: [], // element which appear after animation has finished
       axisCharts: true, // chart type = line or area or bar
       // (refer them also as plot charts in the code)
-      isXNumeric: false, // bool: data was provided in a {[x,y], [x,y]} pattern
       isDataXYZ: false, // bool: data was provided in a {[x,y,z]} pattern
       resized: false, // bool: user has resized
       resizeTimer: null, // timeout function to make a small delay before
       // drawing when user resized
       comboCharts: false, // bool: whether it's a combination of line/column
-      comboChartsHasBars: false, // bool: whether it's a combination of line/column
       dataChanged: false, // bool: has data changed dynamically
       previousPaths: [], // array: when data is changed, it will animate from
       // previous paths
-      seriesXvalues: [], // we will need this in tooltip (it's x position)
-      // when we will have unequal x values, we will need
-      // some way to get x value depending on mouse pointer
-      seriesYvalues: [], // we will need this when deciding which series
-      // user hovered on
-      seriesCandleO: [], // candle stick open values
-      seriesCandleH: [], // candle stick high values
-      seriesCandleL: [], // candle stick low values
-      seriesCandleC: [], // candle stick close values
       allSeriesHasEqualX: true,
-      dataPoints: 0, // the longest series length
       pointsArray: [], // store the points positions here to draw later on hover
       // format is - [[x,y],[x,y]... [x,y]]
       dataLabelsRects: [], // store the positions of datalabels to prevent collision
       lastDrawnDataLabelsIndexes: [],
+      x2SpaceAvailable: 0, // space available on the right side after grid area
       hasNullValues: false, // bool: whether series contains null values
       easing: null, // function: animation effect to apply
       zoomed: false, // whether user has zoomed or not
       gridWidth: 0, // drawable width of actual graphs (series paths)
       gridHeight: 0, // drawable height of actual graphs (series paths)
-      yAxisScale: [],
-      xAxisScale: null,
-      xAxisTicksPositions: [],
-      timescaleTicks: [],
       rotateXLabels: false,
       defaultLabels: false,
       xLabelFormatter: undefined, // formatter for x axis labels
@@ -179,18 +197,16 @@ export default class Globals {
       translateX: 0,
       translateY: 0,
       translateYAxisX: [],
-      yLabelsCoords: [],
-      yTitleCoords: [],
       yAxisWidths: [],
       translateXAxisY: 0,
       translateXAxisX: 0,
-      tooltip: null,
-      tooltipOpts: null
+      tooltip: null
     }
   }
 
   init(config) {
     let globals = this.globalVars(config)
+    this.initGlobalVars(globals)
 
     globals.initialConfig = Utils.extend({}, config)
     globals.initialSeries = JSON.parse(
