@@ -67,20 +67,21 @@ class Intersect {
     let i
     let j
     if (e.target.classList.contains('apexcharts-marker')) {
-      let cx = parseInt(opt.paths.getAttribute('cx'))
-      let cy = parseInt(opt.paths.getAttribute('cy'))
+      let cx = parseInt(opt.paths.getAttribute('cx'), 10)
+      let cy = parseInt(opt.paths.getAttribute('cy'), 10)
       let val = parseFloat(opt.paths.getAttribute('val'))
 
-      j = parseInt(opt.paths.getAttribute('rel'))
+      j = parseInt(opt.paths.getAttribute('rel'), 10)
       i =
         parseInt(
-          opt.paths.parentNode.parentNode.parentNode.getAttribute('rel')
+          opt.paths.parentNode.parentNode.parentNode.getAttribute('rel'),
+          10
         ) - 1
 
       if (ttCtx.intersect) {
         const el = Utils.findAncestor(opt.paths, 'apexcharts-series')
         if (el) {
-          i = parseInt(el.getAttribute('data:realIndex'))
+          i = parseInt(el.getAttribute('data:realIndex'), 10)
         }
       }
 
@@ -128,7 +129,6 @@ class Intersect {
     let bx = 0
     let x = 0
     let y = 0
-    // let bW = 0
     let i = 0
     let strokeWidth
     let barXY = this.getBarTooltipXY({
@@ -143,7 +143,7 @@ class Intersect {
     w.globals.capturedDataPointIndex = j
 
     if (
-      (w.globals.isBarHorizontal && ttCtx.hasBars()) ||
+      (w.globals.isBarHorizontal && ttCtx.tooltipUtil.hasBars()) ||
       !w.config.tooltip.shared
     ) {
       x = barXY.x
@@ -151,7 +151,6 @@ class Intersect {
       strokeWidth = Array.isArray(w.config.stroke.width)
         ? w.config.stroke.width[i]
         : w.config.stroke.width
-      // bW = barXY.barWidth
       bx = x
     } else {
       if (!w.globals.comboCharts && !w.config.tooltip.shared) {
@@ -164,11 +163,10 @@ class Intersect {
       y = w.globals.svgHeight - ttCtx.tooltipRect.ttHeight
     }
 
-    // x exceeds gridWidth
     if (x + ttCtx.tooltipRect.ttWidth > w.globals.gridWidth) {
       x = x - ttCtx.tooltipRect.ttWidth
     } else if (x < 0) {
-      x = x + ttCtx.tooltipRect.ttWidth
+      x = 0
     }
 
     if (ttCtx.w.config.tooltip.followCursor) {
@@ -183,7 +181,7 @@ class Intersect {
     }
 
     if (!w.config.tooltip.shared) {
-      if (w.globals.comboChartsHasBars) {
+      if (w.globals.comboBarCount > 0) {
         ttCtx.tooltipPosition.moveXCrosshairs(bx + strokeWidth / 2)
       } else {
         ttCtx.tooltipPosition.moveXCrosshairs(bx)
@@ -194,22 +192,29 @@ class Intersect {
     if (
       !ttCtx.fixedTooltip &&
       (!w.config.tooltip.shared ||
-        (w.globals.isBarHorizontal && ttCtx.hasBars()))
+        (w.globals.isBarHorizontal && ttCtx.tooltipUtil.hasBars()))
     ) {
-      if (isReversed) {
-        x = w.globals.gridWidth - x
-      }
-      tooltipEl.style.left = x + w.globals.translateX + 'px'
-
-      const seriesIndex = parseInt(
-        opt.paths.parentNode.getAttribute('data:realIndex')
-      )
-
       const isReversed = w.globals.isMultipleYAxis
         ? w.config.yaxis[seriesIndex] && w.config.yaxis[seriesIndex].reversed
         : w.config.yaxis[0].reversed
 
-      if (isReversed && !(w.globals.isBarHorizontal && ttCtx.hasBars())) {
+      if (isReversed) {
+        x = x - ttCtx.tooltipRect.ttWidth
+        if (x < 0) {
+          x = 0
+        }
+      }
+      tooltipEl.style.left = x + w.globals.translateX + 'px'
+
+      const seriesIndex = parseInt(
+        opt.paths.parentNode.getAttribute('data:realIndex'),
+        10
+      )
+
+      if (
+        isReversed &&
+        !(w.globals.isBarHorizontal && ttCtx.tooltipUtil.hasBars())
+      ) {
         y = y + barHeight - (w.globals.series[i][j] < 0 ? barHeight : 0) * 2
       }
       if (ttCtx.tooltipRect.ttHeight + y > w.globals.gridHeight) {
@@ -251,16 +256,19 @@ class Intersect {
       barHeight = barRect.height
       let bw = barRect.width
 
-      let cx = parseInt(bar.getAttribute('cx'))
-      let cy = parseInt(bar.getAttribute('cy'))
+      let cx = parseInt(bar.getAttribute('cx'), 10)
+      let cy = parseInt(bar.getAttribute('cy'), 10)
       barWidth = parseFloat(bar.getAttribute('barWidth'))
       const clientX = e.type === 'touchmove' ? e.touches[0].clientX : e.clientX
 
-      j = parseInt(bar.getAttribute('j'))
-      i = parseInt(bar.parentNode.getAttribute('rel')) - 1
+      j = parseInt(bar.getAttribute('j'), 10)
+      i = parseInt(bar.parentNode.getAttribute('rel'), 10) - 1
+
+      let y1 = bar.getAttribute('data-range-y1')
+      let y2 = bar.getAttribute('data-range-y2')
 
       if (w.globals.comboCharts) {
-        i = parseInt(bar.parentNode.getAttribute('data:realIndex'))
+        i = parseInt(bar.parentNode.getAttribute('data:realIndex'), 10)
       }
 
       // if (w.config.tooltip.shared) {
@@ -275,6 +283,8 @@ class Intersect {
         ttItems: opt.ttItems,
         i,
         j,
+        y1: y1 ? parseInt(y1, 10) : null,
+        y2: y2 ? parseInt(y2, 10) : null,
         shared: ttCtx.showOnIntersect ? false : w.config.tooltip.shared
       })
 
