@@ -81,6 +81,6 @@ describe('Exponential values should parse', () => {
     const maxY = chart.w.globals.maxY
 
     expect(minY).toEqual(2.3e-7)
-    expect(maxY).toEqual(3.2599999999999993e-7)
+    expect(maxY).toEqual(3.0999999999999994e-7)
   })
 })
